añado toda la funcionalidad del carrito de compras, (Controller, Service, Request(que es el model),item, qeu tambien es otro modelo  )

+++ src/main/java/scrum/cannia/model/ItemCarrito.java
@@ -7,7 +7,7 @@ public class ItemCarrito {
 
     private ProductoModel producto;
     private int cantidad;
-    private double subtotal;
+    private int subtotal;
 
     public ItemCarrito(ProductoModel producto, int cantidad) {
         this.producto = producto;
@@ -15,9 +15,6 @@ public class ItemCarrito {
         calcularSubtotal();
     }
 
-//     ==========================================
-//     METODO NECESARIO PARA QUE TODO FUNCIONE
-//     ==========================================
     public void calcularSubtotal() {
         if (producto != null) {
             this.subtotal = producto.getValor() * cantidad;
